devfreq_boost: Introduce devfreq boost driver
This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are
near-zero at the boosted frequencies and power consumption is minimized
at said frequencies. The goal of this driver is to provide an interface
to achieve optimal device performance by requesting boosts on key
events, such as when a frame is ready to rendered to the display.

Supported devfreq devices: LLCCBW_DDR, CPU_LLCCBW, GPUBW

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[@ 0ctobot: msm-4.14: Adapted for compatibility with drm_notifier]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: hypeartist <hypeartist@gmail.com>